+++ praetorian_cli/handlers/cli_decorators.py
@@ -26,12 +26,13 @@ def cli_handler(func):
     return func
 
 
-def list_params(filter_by):
+def list_params(filter_by, has_details=True):
     def decorator(func):
         func = click.option('-f', '--filter', default='', help=f'Filter by {filter_by}')(func)
-        func = click.option('-d', '--details', is_flag=True, default=False, help='Show detailed information')(func)
         func = pagination(func)
         func = cli_handler(func)
+        if has_details:
+            func = click.option('-d', '--details', is_flag=True, default=False, help='Show detailed information')(func)
         return func
 
     return decorator