+++ praetorian_cli/handlers/unlink.py
@@ -14,5 +14,14 @@ def unlink():
 @cli_handler
 @click.argument('username')
 def account(chariot, username):
-    """ Remove a collaborator account from your account """
+    """ Remove a collaborator account from your account. This will
+    revoke their access to your account.
+
+    Arguments:
+        - NAME: Their email address.
+
+    \b
+    Example usages:
+        - praetorian chariot unlink account john@praetorian.com
+    """
     chariot.accounts.delete_collaborator(username)